Fix env_creator tool not to raise error when there's no VScode launch.json template file

+++ env_collection/env_creator.py
@@ -18,8 +18,13 @@ def json_to_envfiles(output_file: pathlib.Path):
         input_env_file_content: str = output_file.read_text()
         input_env: dict[str, typing.Union[str, dict[str, str]]] = json.loads(input_env_file_content)
 
-        template_launch_json_file: pathlib.Path = pathlib.Path('template/launch.json')
-        template_launch_json_file_content: dict[str, object] = json.loads(template_launch_json_file.read_text())
+        template_launch_json_file: typing.Optional[pathlib.Path] = pathlib.Path('template/launch.json')
+        if not template_launch_json_file.exists():
+            print('There\'s no launch.json template file for VS Code.\n'
+                  'Place file on template/launch.json')
+            template_launch_json_file = None
+        else:
+            template_launch_json_file_content: dict[str, object] = json.loads(template_launch_json_file.read_text())
 
         output_docker_file: pathlib.Path = pathlib.Path(output_name+'.env')
         output_bash_file: pathlib.Path = pathlib.Path(output_name+'.sh')
@@ -35,10 +40,11 @@ def json_to_envfiles(output_file: pathlib.Path):
         if output_ps_file.exists():
             output_ps_file.unlink()
 
-        if not output_launch_json_file.parent.exists():
-            output_launch_json_file.parent.mkdir()
-        if output_launch_json_file.exists():
-            output_launch_json_file.unlink()
+        if template_launch_json_file is not None:
+            if not output_launch_json_file.parent.exists():
+                output_launch_json_file.parent.mkdir()
+            if output_launch_json_file.exists():
+                output_launch_json_file.unlink()
 
         with output_docker_file.open('w') as docker_fp,\
              output_bash_file.open('w') as bash_fp,\
@@ -78,18 +84,19 @@ def json_to_envfiles(output_file: pathlib.Path):
                 bash_fp.write(bash_line)
                 ps_fp.write(ps_line)
 
-        with output_launch_json_file.open('w') as launch_json_fp:
-            launch_json_env = dict()
-            for k, v in input_env.items():
-                if k.startswith('__comment') or k.startswith('__line_break'):
-                    continue
-                elif type(v) == str:
-                    launch_json_env[k] = v
-                else:
-                    launch_json_env[k] = v['vscode_launch'].format(**input_env)
+        if template_launch_json_file is not None:
+            with output_launch_json_file.open('w') as launch_json_fp:
+                launch_json_env = dict()
+                for k, v in input_env.items():
+                    if k.startswith('__comment') or k.startswith('__line_break'):
+                        continue
+                    elif type(v) == str:
+                        launch_json_env[k] = v
+                    else:
+                        launch_json_env[k] = v['vscode_launch'].format(**input_env)
 
-            template_launch_json_file_content['configurations'][0]['env'] = launch_json_env
-            launch_json_fp.write(json.dumps(template_launch_json_file_content, indent=4))
+                template_launch_json_file_content['configurations'][0]['env'] = launch_json_env
+                launch_json_fp.write(json.dumps(template_launch_json_file_content, indent=4))
     except Exception as e:
         print('Exception raised!')
         print(get_traceback_msg(e))
